Move FrameRenderer into Fast Renderers
