+++ Xamarin.Forms.Platform.Android/Xamarin.Forms.Platform.Android.csproj
@@ -100,10 +100,10 @@
     </Compile>
     <Compile Include="AndroidApplicationLifecycleState.cs" />
     <Compile Include="AndroidTitleBarVisibility.cs" />
-    <Compile Include="AppCompat\FastButtonRenderer.cs" />
+    <Compile Include="FastRenderers\ButtonRenderer.cs" />
     <Compile Include="AppCompat\FormsViewPager.cs" />
     <Compile Include="AppCompat\FragmentContainer.cs" />
-    <Compile Include="AppCompat\FrameRenderer.cs" />
+    <Compile Include="FastRenderers\FrameRenderer.cs" />
     <Compile Include="AppCompat\IManageFragments.cs" />
     <Compile Include="AppCompat\MasterDetailContainer.cs" />
     <Compile Include="AppCompat\Platform.cs" />
